Make manifesto the default landing page
